requirements.txt
# RSS抓取相关
feedparser==6.0.10
requests==2.31.0

# AI处理相关
openai>=1.0.0

# HTML生成相关
jinja2==3.1.3

# 定时任务
schedule==1.2.0

# 工具库
python-dateutil==2.8.2
pytz==2023.3

# 日志
colorlog==6.7.0

# 环境变量
python-dotenv==1.0.0
